[polish] separate inliner, linter, and extractor
